File I/O

+++ 0x15-file_io/0-read_textfile.c
@@ -12,7 +12,8 @@
 
 ssize_t read_textfile(const char *filename, size_t letters)
 {
-	ssize_t bytesRead, bytesWritten;
+	int bytesRead, bytesWritten;
+	
 	FILE *file = fopen(filename, "r");
 	char *buffer = (char *)malloc(letters + 1);
